qnamaker mvc create kb with solution
