+++ quickstarts/create-bk/QnaMakerQuickstart.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27703.2042
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "QnaMakerQuickstart", "QnaMakerQuickstart\QnaMakerQuickstart.csproj", "{3846A675-5ACE-4407-B98C-8EA33157B6BD}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3846A675-5ACE-4407-B98C-8EA33157B6BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3846A675-5ACE-4407-B98C-8EA33157B6BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3846A675-5ACE-4407-B98C-8EA33157B6BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3846A675-5ACE-4407-B98C-8EA33157B6BD}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {6F39C823-6C2D-4415-89A4-7F11958F8A64}
+	EndGlobalSection
+EndGlobal
